+++ lib/pg_export/lib/pg_export/factories/ftp_gateway_factory.rb
@@ -2,16 +2,16 @@
 
 # auto_register: false
 
-require 'pg_export/lib/pg_export/adapters/ftp_adapter'
+require 'pg_export/lib/pg_export/gateways/ftp'
 require 'pg_export/import'
 
 class PgExport
   module Factories
-    class FtpAdapterFactory
+    class FtpGatewayFactory
       include Import['config']
 
-      def ftp_adapter
-        ::PgExport::Adapters::FtpAdapter.new(
+      def ftp_gateway
+        ::PgExport::Gateways::Ftp.new(
           host: config.ftp_host,
           user: config.ftp_user,
           password: config.ftp_password