+++ lib/pg_export/lib/pg_export/listeners/interactive/upload_dump_to_ftp.rb
@@ -7,7 +7,7 @@ class PgExport
     class Interactive
       class UploadDumpToFtp < InteractiveListener
         def on_step(step_name:, args:)
-          @spinner = build_spinner("Uploading #{args.first[:dump]} to #{args.first[:ftp_adapter]}")
+          @spinner = build_spinner("Uploading #{args.first[:dump]} to #{args.first[:ftp_gateway]}")
         end
 
         def on_step_succeeded(*)